поправлены проблемы с компиляцией под линукс. полностью реализован механизм билда статичечской и динамической библиотеки
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@63666 954022d7-b5bf-4e40-9824-e11837661b57

+++ PdfWriter/PdfRenderer.h
@@ -306,7 +306,7 @@ private:
 		}
 		inline void   SetAlpha(const LONG& lAlpha)
 		{
-			m_nAlpha = (BYTE)max(0, min(255, lAlpha));
+			m_nAlpha = (BYTE)max(0, min(255, (int)lAlpha));
 		}
 		inline double GetSize()
 		{
@@ -536,7 +536,7 @@ private:
 		}
 		inline void         SetAlpha1(const LONG& lAlpha)
 		{
-			m_nAlpha1 = (BYTE)max(0, min(255, lAlpha));
+			m_nAlpha1 = (BYTE)max(0, min(255, (int)lAlpha));
 		}
 		inline LONG         GetAlpha2()
 		{
@@ -544,7 +544,7 @@ private:
 		}
 		inline void         SetAlpha2(const LONG& lAlpha)
 		{
-			m_nAlpha2 = (BYTE)max(0, min(255, lAlpha));
+			m_nAlpha2 = (BYTE)max(0, min(255, (int)lAlpha));
 		}
 		inline std::wstring GetTexturePath()
 		{
@@ -568,7 +568,7 @@ private:
 		}
 		inline void         SetTextureAlpha(const LONG& lAlpha)
 		{
-			m_nTextureAlpha = (BYTE)max(0, min(255, lAlpha));
+			m_nTextureAlpha = (BYTE)max(0, min(255, (int)lAlpha));
 		}
 		inline double       GetLinearAngle()
 		{
@@ -863,10 +863,10 @@ private:
 				TColor oColor1 = lColor1;
 				TColor oColor2 = lColor2;
 
-				BYTE r = (BYTE)max(0, min(255, (oColor1.r + (oColor2.r - oColor1.r) / dDiff * (dDstPoint - dPoint1))));
-				BYTE g = (BYTE)max(0, min(255, (oColor1.g + (oColor2.g - oColor1.g) / dDiff * (dDstPoint - dPoint1))));
-				BYTE b = (BYTE)max(0, min(255, (oColor1.b + (oColor2.b - oColor1.b) / dDiff * (dDstPoint - dPoint1))));
-				BYTE a = (BYTE)max(0, min(255, (oColor1.a + (oColor2.a - oColor1.a) / dDiff * (dDstPoint - dPoint1))));
+				BYTE r = (BYTE)max(0, min(255, (int)(oColor1.r + (oColor2.r - oColor1.r) / dDiff * (dDstPoint - dPoint1))));
+				BYTE g = (BYTE)max(0, min(255, (int)(oColor1.g + (oColor2.g - oColor1.g) / dDiff * (dDstPoint - dPoint1))));
+				BYTE b = (BYTE)max(0, min(255, (int)(oColor1.b + (oColor2.b - oColor1.b) / dDiff * (dDstPoint - dPoint1))));
+				BYTE a = (BYTE)max(0, min(255, (int)(oColor1.a + (oColor2.a - oColor1.a) / dDiff * (dDstPoint - dPoint1))));
 
 				TColor oResColor;
 				oResColor.Set(r, g, b, a);
